test: inline shift+click helper into multiselect test for TDD base-branch compatibility
